update: add computer and human choice generation
This adds random choice of rock,paper,scissors
for the computer opponent and a manual input
for the human player. The input is validated:
empty and undefined entries, and everything that
is not rock, paper or scissors, will prompt for
the input again until its correct. The general
message that lets the player know that something
is wrong with the input is displayed as a prompt
text.

There are also variables to store computer and
human scores, for now they are just defaulted to
0.