update & fix: add ability to play one round
Added playRound(), which checks the inputs
and prints out the result (win, lose or draw) with
the new printRoundResult() function.

Made the error message in getHumanChoice() into a
separate variable, rather than overwriting the
"message" variable.

Added formatHumanChoice() function to convert
humanInput to lowercase and then to capitalize the
first letter it for easier printing and
comparisons.

Fixed validateHumanInput() when checking if the
input is null, previously didn't check for null
at all.