update: Add the playGame function
Moved the score declaration and playRound() into
the new playGame() function. It loops for 5
rounds. Prints out the results of each round and
the overall game result (win, lose or draw).

Added the string return in playRound(), so it
returns the string result of the round ("win",
"loss" or "draw"), which is used by the
switch-case, to determine what message to print
and what score to increment.